+++ packages/ranges-sort/es5/index.js
@@ -20,7 +20,7 @@ var isArr = Array.isArray;
 // does this: [ [2, 5], [1, 6] ] => [ [1, 6], [2, 5] ]
 // sorts first by first element, then by second. Retains possible third element.
 
-function srt(arrOfRanges, o) {
+function srt(arrOfRanges, originalOptions) {
   // arrOfRanges validation
   if (!isArr(arrOfRanges)) {
     throw new TypeError('ranges-sort: [THROW_ID_01] Input must be an array, consisting of range arrays! Currently its type is: ' + (typeof arrOfRanges === 'undefined' ? 'undefined' : _typeof(arrOfRanges)) + ', equal to: ' + JSON.stringify(arrOfRanges, null, 4));
@@ -35,7 +35,7 @@ function srt(arrOfRanges, o) {
   var defaults = {
     strictlyTwoElementsInRangeArrays: false
     // fill any settings with defaults if missing:
-  };var opts = Object.assign({}, defaults, o);
+  };var opts = Object.assign({}, defaults, originalOptions);
   // the check:
   checkTypes(opts, defaults, { msg: 'ranges-sort: [THROW_ID_02*]' });