+++ packages/ranges-sort/changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2017-09-13
+### Added
+- ✨ More description in readme.
+- ✨ More unit tests, including tests for examples used in readme.
+
 ## [2.0.0] - 2017-09-12
 ### Added
 - ✨ `opts.strictlyTwoElementsInRangeArrays` (default is `false`, differently from v1 which is opposite).
@@ -12,4 +17,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - ✨ Initial release
 
+[2.1.0]: https://github.com/codsen/ranges-sort/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/codsen/ranges-sort/compare/v1.0.0...v2.0.0